+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-migrate/cdk-migrate--from-stack-creates-deployable-app-typescript.integtest.ts
@@ -1,0 +1,11 @@
+import { fromStackCreatesDeployableApp } from "./testcase";
+import { integTest, withExtendedTimeoutFixture } from "../../../lib";
+
+const language = 'typescript';
+
+integTest(
+  `cdk migrate --from-stack creates deployable typescript app`,
+  withExtendedTimeoutFixture(async (fixture) => {
+    await fromStackCreatesDeployableApp(fixture, language);
+  }),
+);
